fixed mod to work with multi-dimensions/surfaces
entities are keyed off of a magic number between 1 and 5 + the chests x and y coordinates. Thus, if the magic number picked for your chest in your factorissimo dimension was 1 and the x and y coords were 1 and 1, then the next time you put a chest down in the same place in a different factorissimo dimension (and that magic number is also 1) then only one of them will work as it will bump the other one out of the update queue. This was a pain in the ass to figure out but I have fixed it with a simple patch.

The patch modifies the keying system to also take in account the surface/dimension index, this way it is literally impossible to overwrite the previous chests as they'd have to be placed in the same x and y AND in the same factorissimo dimension or any other surface.